exception and event

+++ src/main/java/com/yellowstone/cleantogather/api/event/EventRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.repository.CrudRepository;
 
 // This will be auto implemented by Spring
 // Crud refers to Create Read Update Delete
-public interface EventRepository extends CrudRepository<Event, Integer> {
+public interface EventRepository extends CrudRepository<Event, Long> {
 }
